requirements.txt

Django==5.2.6
django-cors-headers==4.5.0
django-deep-serializer==0.1.3
django-filter==25.2
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
psycopg2==2.9.10
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1


